Fixed write_files ownership bug
Write_files happens before the users and groups are created. Because of that,
it would error out on the setting ownership step. Also, writing to the user's
home dir before it exists means the user doesn't get .bashrc and .profile files
in their home dir. Instead, we write to a temp dir and move the file later with
runcmd